requirements.txt
flask
web3
requests
python-dotenv
matplotlib
pandas
scikit-learn
